move robot_Recovery code to an external file
